Refactor merged branches Add DishCategories controller, routes, refactored DishesControllers/Dish.create method, ProductsControllers/Product.deleteById, and other controllers; fix migrations, fix model relationships and models themselves; add todos to routes, rename some routes

+++ apps/backend/app/Controllers/Http/ProductsControllers/Product.ts
@@ -1,4 +1,5 @@
 import type { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
+import ActionForbiddenException from 'App/Exceptions/Auth/ActionForbiddenException';
 import ProductModel from 'App/Models/Product';
 import ProductCategory from 'App/Models/ProductCategory';
 import CreateProductValidator from 'App/Validators/Product/CreateProductValidator';
@@ -11,7 +12,6 @@ export default class ProductController {
     |
     | description: Creates new product.
     | middlewares: AuthMiddleware
-    | exceptions: AuthenticationException("E_UNAUTHORIZED_ACCESS"), 
     | returns: ProductModule
     */
     public async create({ request, auth } : HttpContextContract){
@@ -37,9 +37,25 @@ export default class ProductController {
         return await product.save();
     }
 
-    public async deleteById({ request }: HttpContextContract){
-        const productId = request.input('id');
-        const product = await ProductModel.findOrFail(productId);
-        await product.delete()
+
+    /*
+    |--------------------------------------------------------------------------
+    | DELETE /products/:id
+    |--------------------------------------------------------------------------
+    |
+    | description: Deletes product with specified id.
+    | middlewares: AuthMiddleware
+    | returns: ProductModule
+    */
+    public async deleteById({ params, auth }: HttpContextContract){
+        const product = await ProductModel.findOrFail(params.id);
+
+        // Checking if current user is the author of this product
+        if (product.authorId !== auth.user!.id) {
+            throw new ActionForbiddenException('You are not the author of this product');
+        };
+
+        await product.delete();
+        return product.toJSON();
     }
 }
